% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrsensemakr.R
\name{mr_sensemakr}
\alias{mr_sensemakr}
\title{Sensitivity Analysis for Mendelian Randomization}
\usage{
mr_sensemakr(
  outcome,
  exposure,
  instrument,
  covariates = NULL,
  data,
  benchmark_covariates = NULL,
  k = 1,
  alpha = 0.05
)
}
\arguments{
\item{outcome}{A character vector with the name of the outcome trait.}

\item{exposure}{A character vector with the name of the exposure trait.}

\item{instrument}{A character vector with the name of the genetic instrument (e.g., a polygenic risk score).}

\item{covariates}{A character vector with the name of the control covariates, such as age, sex, genomic principal components, batch effect dummies and putative pleiotropic pathways.}

\item{data}{An object of the class data.frame containing the variables used in the analysis.}

\item{benchmark_covariates}{Covariates for benchmarking. Must be a subset of the \code{covariates} argument. The user has two options: (i) character vector of the names of covariates that will be used to bound the plausible strength of the unmeasured variables. Each variable will be considered separately; (ii) a named list with character vector names of covariates that will be used, as a group, to bound the plausible strength of the unmeasured variables. The names of the list will be used for the benchmark labels.}

\item{k}{numeric vector. Parameterizes how many times stronger the unmeasured variables are, in explaining the genetic instrument and the exposure/outcome traits, in comparison to the observed benchmark covariates.}

\item{alpha}{significance level.}
}
\value{
An object of class \code{mr_sensemakr}, which is a list containing:
\describe{
  \item{info}{A list with the outcome, exposure, instrument, covariates, alpha level, and missing data information.}
  \item{exposure}{A list with the regression model for the genetic association with the exposure, along with its sensitivity statistics and (if benchmarks provided) bounds on the bias.}
  \item{outcome}{A list with the regression model for the genetic association with the outcome, along with its sensitivity statistics and (if benchmarks provided) bounds on the bias.}
  \item{mr}{An object of class \code{trad.mr} with the traditional MR (2SLS) estimates.}
}
}
\description{
Performs sensitivity analysis for Mendelian Randomization (MR) studies.
The method quantifies the robustness of MR findings to potential violations
of the MR assumptions due to residual population stratification, batch effects
and horizontal pleiotropy. It computes sensitivity statistics---the partial
R-squared of the genetic instrument with the exposure and outcome traits, and
the robustness value---that reveal the minimal strength that unmeasured
variables \code{W} would need to have to explain away the MR results.
}
\examples{
## loads package
library(mrsensemakr)

## simulated data example
data("sim_data")

## create vectors indicating variable names in the data
outcome    <- "out.trait" # name of outcome trait
exposure   <- "exp.trait" # name of exposure trait
instrument <- "prs" # genetic instrument (e.g, polygenic risk score)
age.sex    <- c("age", "sex") # age and sex variables (if applicable)
alc.smok   <- c("alcohol", "smoking") # putative pleoitropic vars.
pcs        <- paste0("pc", 1:20) # first 20 principal components pc1 ... pc20

## runs MR sensitivity analysis
mr.sense <- mr_sensemakr(outcome = outcome,
                         exposure = exposure,
                         instrument = instrument,
                         covariates = c(age.sex, alc.smok, pcs),
                         data = sim_data,
                         benchmark_covariates = list(alc.smok = alc.smok,
                                                     pcs = pcs))
## print results
mr.sense


## sensitivity contour plots
plot(mr.sense,
     benchmark_covariates = list(alc.smok = alc.smok,
                                 pcs = pcs),
     k = list(alc.smok = 25,
              pcs = 35))
}
