% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrsensemakr.R
\name{mr_sensemakr}
\alias{mr_sensemakr}
\title{MR-sensemakr}
\usage{
mr_sensemakr(
  outcome,
  exposure,
  instrument,
  covariates = NULL,
  data,
  benchmark_covariates = NULL,
  k = 1,
  alpha = 0.05
)
}
\description{
MR-sensemakr
}
