% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrsensemakr.R
\name{mr_sensemakr}
\alias{mr_sensemakr}
\title{MR-sensemakr}
\usage{
mr_sensemakr(
  outcome,
  exposure,
  instrument,
  covariates = NULL,
  data,
  benchmark_covariates = NULL,
  k = 1,
  alpha = 0.05
)
}
\arguments{
\item{outcome}{A character vector with the name of the outcome trait.}

\item{exposure}{A character vector with the name of the exposure trait.}

\item{instrument}{A character vector with the name of the genetic instrument.}

\item{covariates}{A character vector with the name of the control covariates, such as age, sex, genomic principal components, batch effect dummies and putative pleiotropic pathways.}

\item{data}{An object of the class data.frame containing the variables used in the analysis.}

\item{benchmark_covariates}{The user has two options: (i) character vector of the names of covariates that will be used to bound the plausible strength of the unobserved confounders. Each variable will be considered separately; (ii) a named list with character vector names of covariates that will be used, as a group, to bound the plausible strength of the unobserved confounders. The names of the list will be used for the benchmark labels.}

\item{k}{numeric vector. Parameterizes how many times stronger residual biases are related to the treatment and the outcome in comparison to the observed benchmark covariates.}

\item{alpha}{significance level}
}
\description{
MR-sensemakr
}
