% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrsensemakr.R
\name{print.mr_sensemakr}
\alias{print.mr_sensemakr}
\title{Print results of MR sensitivity analysis}
\usage{
\method{print}{mr_sensemakr}(x, digits = 2, ...)
}
\arguments{
\item{x}{an object of class \code{mr_sensemakr}.}

\item{digits}{number of digits to round results to.}

\item{...}{additional arguments passed to other print methods.}
}
\value{
The function is called for its side effect of printing to the console.
  Invisibly returns the input \code{x}.
}
\description{
Prints a summary of the Mendelian Randomization sensitivity analysis,
including traditional MR estimates, sensitivity statistics for the
exposure and outcome regressions, and (if available) benchmark bounds.
}
